registry: one parser for portolan.json, shared with what sync will need
